+++ lib/Backend/triggers/sensor_definition_action_definition.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/widgets.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'sensor_definition_action_definition.freezed.dart';
@@ -8,6 +9,7 @@ abstract class TriggerActionDef with _$TriggerActionDef {
   const factory TriggerActionDef({
     required String name,
     required Function translated,
+    Widget? icon,
     required String uuid,
     @Default(false) final bool defaultActions,
   }) = _TriggerActionDef;